+++ docs/CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## v1.0.0 -- 2022-07-03
+
+### Added or Changed
+* Add fighter entry to the database.
+* Add driver entry to the database.
+* Delete drivers and fighters from the database.
+* Optimal lineup page setup - needs more logic
+
 ## v1.0.0 -- 2022-06-24
 
 ### Added or Changed
 * Setup main UI for the application
-* 
 
 ### Removed
-*